pyproject.toml
[tool.poetry]
name = "face-recognition-system"
version = "0.1.0"
description = ""
authors = ["adpakw"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ipykernel = "^6.29.5"
torch = "^2.5.1"
torchvision = "^0.20.1"
opencv-python = "^4.10.0.84"
scipy = "^1.14.1"
tqdm = "^4.67.1"
ipywidgets = "^8.1.5"
ffmpeg-python = "^0.2.0"
requests = "^2.32.3"
scikit-image = "^0.25.2"
pandas = "^2.2.3"
matplotlib = "^3.10.1"
opencv-python-headless = "^4.11.0.86"
opencv-contrib-python = "^4.11.0.86"
pydantic = "^2.11.3"
pyyaml = "^6.0.2"
torchfile = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
